+++ docs/CHANGELOG.md
@@ -50,6 +50,7 @@
 - added new creature explosions effects
 - added meteorite artifacts support
 - added examine item feature for certain items
+- added Monkey control
 - added Tony control
 - added Spikes animation in Coastal Village and Madubu Gorge
 - added Aldwych Drill control (Spike Ceiling with timer=1 to descend faster)